+++ src/axios-cached-dns-resolve.ts
@@ -1,22 +1,24 @@
-/* eslint-disable no-plusplus */
 import dns from 'dns'
 import URL from 'url'
 import net from 'net'
 import stringify from 'json-stringify-safe'
 import LRUCache from 'lru-cache'
 import util from 'util'
-import { init as initLogger } from './logging.js'
+import { init as initLogger } from './logging'
+import { Logger } from 'pino'
+import { AxiosInstance } from 'axios'
+import { Config, CacheConfig, Stats, DnsEntry } from "./index.d";
 
 const dnsResolve = util.promisify(dns.resolve)
 const dnsLookup = util.promisify(dns.lookup)
 
 export const config = {
   disabled: process.env.AXIOS_DNS_DISABLE === 'true',
-  dnsTtlMs: process.env.AXIOS_DNS_CACHE_TTL_MS || 5000, // when to refresh actively used dns entries (5 sec)
-  cacheGraceExpireMultiplier: process.env.AXIOS_DNS_CACHE_EXPIRE_MULTIPLIER || 2, // maximum grace to use entry beyond TTL
-  dnsIdleTtlMs: process.env.AXIOS_DNS_CACHE_IDLE_TTL_MS || 1000 * 60 * 60, // when to remove entry entirely if not being used (1 hour)
-  backgroundScanMs: process.env.AXIOS_DNS_BACKGROUND_SCAN_MS || 2400, // how frequently to scan for expired TTL and refresh (2.4 sec)
-  dnsCacheSize: process.env.AXIOS_DNS_CACHE_SIZE || 100, // maximum number of entries to keep in cache
+  dnsTtlMs: parseInt(process.env.AXIOS_DNS_CACHE_TTL_MS || '5000'), // when to refresh actively used dns entries (5 sec)
+  cacheGraceExpireMultiplier: parseInt(process.env.AXIOS_DNS_CACHE_EXPIRE_MULTIPLIER || '2'), // maximum grace to use entry beyond TTL
+  dnsIdleTtlMs: parseInt(process.env.AXIOS_DNS_CACHE_IDLE_TTL_MS || '1000') * 60 * 60, // when to remove entry entirely if not being used (1 hour)
+  backgroundScanMs: parseInt(process.env.AXIOS_DNS_BACKGROUND_SCAN_MS || '2400'), // how frequently to scan for expired TTL and refresh (2.4 sec)
+  dnsCacheSize: parseInt(process.env.AXIOS_DNS_CACHE_SIZE || '100'), // maximum number of entries to keep in cache
   // pino logging options
   logging: {
     name: 'axios-cache-dns-resolve',
@@ -25,18 +27,18 @@ export const config = {
     // timestamp: true,
     prettyPrint: process.env.NODE_ENV === 'DEBUG' || false,
     formatters: {
-      level(label/* , number */) {
+      level(label: string, number: number) {
         return { level: label }
       },
     },
   },
-  cache: undefined,
-}
+  cache: undefined as LRUCache<string, any> | undefined,
+} as Config;
 
 export const cacheConfig = {
   max: config.dnsCacheSize,
   ttl: (config.dnsTtlMs * config.cacheGraceExpireMultiplier), // grace for refresh
-}
+} as CacheConfig;
 
 export const stats = {
   dnsEntries: 0,
@@ -47,11 +49,11 @@ export const stats = {
   errors: 0,
   lastError: 0,
   lastErrorTs: 0,
-}
+} as Stats;
 
-let log
-let backgroundRefreshId
-let cachePruneId
+let log: Logger
+let backgroundRefreshId: NodeJS.Timeout
+let cachePruneId: NodeJS.Timeout
 
 init()
 
@@ -60,16 +62,21 @@ export function init() {
 
   if (config.cache) return
 
-  config.cache = new LRUCache(cacheConfig)
+  config.cache = new LRUCache<string, any>(cacheConfig)
 
   startBackgroundRefresh()
   startPeriodicCachePrune()
-  cachePruneId = setInterval(() => config.cache.purgeStale(), config.dnsIdleTtlMs)
 }
 
 export function reset() {
-  if (backgroundRefreshId) clearInterval(backgroundRefreshId)
-  if (cachePruneId) clearInterval(cachePruneId)
+  if (backgroundRefreshId) {
+    clearInterval(backgroundRefreshId)
+    backgroundRefreshId.unref();
+  }
+  if (cachePruneId) {
+    clearInterval(cachePruneId)
+    cachePruneId.unref();
+  }
 }
 
 export function startBackgroundRefresh() {
@@ -79,16 +86,16 @@ export function startBackgroundRefresh() {
 
 export function startPeriodicCachePrune() {
   if (cachePruneId) clearInterval(cachePruneId)
-  cachePruneId = setInterval(() => config.cache.purgeStale(), config.dnsIdleTtlMs)
+  cachePruneId = setInterval(() => config.cache?.purgeStale(), config.dnsIdleTtlMs)
 }
 
 export function getStats() {
-  stats.dnsEntries = config.cache.size
+  stats.dnsEntries = config.cache?.size || 0
   return stats
 }
 
 export function getDnsCacheEntries() {
-  return Array.from(config.cache.values())
+  return Array.from(config.cache?.values() || [])
 }
 
 // const dnsEntry = {
@@ -103,30 +110,34 @@ export function getDnsCacheEntries() {
 //   updatedTs: 1555771516581,
 // }
 
-export function registerInterceptor(axios) {
+export function registerInterceptor(axios: AxiosInstance) {
   if (config.disabled || !axios || !axios.interceptors) return // supertest
   axios.interceptors.request.use(async (reqConfig) => {
     try {
+      if (!reqConfig.headers) {
+        reqConfig.headers = {};
+      }
+
       let url
       if (reqConfig.baseURL) {
         url = URL.parse(reqConfig.baseURL)
       } else {
-        url = URL.parse(reqConfig.url)
+        url = URL.parse(reqConfig.url || '')
       }
 
-      if (net.isIP(url.hostname)) return reqConfig // skip
+      if (net.isIP(url.hostname || '')) return reqConfig // skip
 
-      reqConfig.headers.Host = url.hostname // set hostname in header
+      reqConfig.headers.Host = url.hostname || '' // set hostname in header
 
-      url.hostname = await getAddress(url.hostname)
-      delete url.host // clear hostname
+      url.hostname = await getAddress(url.hostname || '')
+      url.host = null // clear hostname
 
       if (reqConfig.baseURL) {
         reqConfig.baseURL = URL.format(url)
       } else {
         reqConfig.url = URL.format(url)
       }
-    } catch (err) {
+    } catch (err: any) {
       recordError(err, `Error getAddress, ${err.message}`)
     }
 
@@ -134,17 +145,17 @@ export function registerInterceptor(axios) {
   })
 }
 
-export async function getAddress(host) {
-  let dnsEntry = config.cache.get(host)
+export async function getAddress(host: string) {
+  let dnsEntry = config.cache?.get(host)
   if (dnsEntry) {
-    ++stats.hits
+    stats.hits += 1
     dnsEntry.lastUsedTs = Date.now()
     // eslint-disable-next-line no-plusplus
     const ip = dnsEntry.ips[dnsEntry.nextIdx++ % dnsEntry.ips.length] // round-robin
-    config.cache.set(host, dnsEntry)
+    config.cache?.set(host, dnsEntry)
     return ip
   }
-  ++stats.misses
+  stats.misses += 1
   if (log.isLevelEnabled('debug')) log.debug(`cache miss ${host}`)
 
   const ips = await resolve(host)
@@ -154,10 +165,10 @@ export async function getAddress(host) {
     nextIdx: 0,
     lastUsedTs: Date.now(),
     updatedTs: Date.now(),
-  }
+  } as DnsEntry
   // eslint-disable-next-line no-plusplus
   const ip = dnsEntry.ips[dnsEntry.nextIdx++ % dnsEntry.ips.length] // round-robin
-  config.cache.set(host, dnsEntry)
+  config.cache?.set(host, dnsEntry)
   return ip
 }
 
@@ -166,28 +177,28 @@ export async function backgroundRefresh() {
   if (backgroundRefreshing) return // don't start again if currently iterating slowly
   backgroundRefreshing = true
   try {
-    config.cache.forEach(async (value, key) => {
+    config.cache?.forEach(async (value, key) => {
       try {
         if (value.updatedTs + config.dnsTtlMs > Date.now()) {
           return // continue/skip
         }
         if (value.lastUsedTs + config.dnsIdleTtlMs <= Date.now()) {
-          ++stats.idleExpired
-          config.cache.delete(key)
+          stats.idleExpired += 1
+          config.cache?.delete(key)
           return // continue
         }
 
         const ips = await resolve(value.host)
         value.ips = ips
         value.updatedTs = Date.now()
-        config.cache.set(key, value)
-        ++stats.refreshed
-      } catch (err) {
+        config.cache?.set(key, value)
+        stats.refreshed += 1
+      } catch (err: any) {
         // best effort
         recordError(err, `Error backgroundRefresh host: ${key}, ${stringify(value)}, ${err.message}`)
       }
     })
-  } catch (err) {
+  } catch (err: any) {
     // best effort
     recordError(err, `Error backgroundRefresh, ${err.message}`)
   } finally {
@@ -200,7 +211,7 @@ export async function backgroundRefresh() {
  * @param host
  * @returns {*[]}
  */
-async function resolve(host) {
+async function resolve(host: string) {
   let ips
   try {
     ips = await dnsResolve(host)
@@ -217,15 +228,14 @@ async function resolve(host) {
 // ***************** { address: '142.250.190.68', family: 4 }
 // , { all: true } /***************** [ { address: '142.250.190.68', family: 4 } ]
 
-function extractAddresses(lookupResp) {
+function extractAddresses(lookupResp: any) {
   if (!Array.isArray(lookupResp)) throw new Error('lookup response did not contain array of addresses')
   return lookupResp.filter((e) => e.address != null).map((e) => e.address)
 }
 
-function recordError(err, errMesg) {
-  ++stats.errors
+function recordError(err: any, errMesg: string) {
+  stats.errors += 1
   stats.lastError = err
-  stats.lastErrorTs = new Date().toISOString()
+  stats.lastErrorTs = Date.now()
   log.error(err, errMesg)
 }
-/* eslint-enable no-plusplus */